cpufreq: add global var to track current max freq for deep sleep patch
add a global variable to cpufreq.c for tracking current max frequency.
this allows the deep sleep patch to properly resume at the max frequency
from before the device went to sleep, instead of automatically jumping
back to an arbitrary frequency, or back to the highest frequency available.

this commit allows users to change their max frequency setting, and still
allow the deep sleep patch to detect the correct frequency.